Make QueueViewActivity singleTask to allow implementing navigation
This is to allow a hack that implements closing the tab from one of
our own actions, or even navigation. We can't explicitly close the custom
tab, and we also can't tell it to navigate to a different page. However
we can use an Intent to start the underlying QueueViewActivity - with
singleTask this simply results in the custom tab being removed, meaning
we end up with only the QueueViewActivity on the stack.

This may or may not cause us other issues down the line - however given
that the page list (QueueViewActivity) is conceptually the root of any
other user interaction, this is probably not a bad choice.a

(The actual Activity navigation and intents will be in a separate commit.)