Make the search bar non-scrolling
RecyclerView really doesn't like not being able to scroll. Using
heigh=wrap_content and disabling nested scrolling results in the
RecycleView breaking (it only shows as many items as would fit on
screen). Instead lets just keep the search bar fixed, and only
scroll the RecyclerView. We can fix this in future if needed, but
this isn't really important for the purposes of prototyping.

This of course is terribly annoying on small devices, but who uses small
phones nowadays anyways? </sarcasm>